requirements.txt
Django==2.0.5
pytz==2018.4
httplib2==0.11.3
feedparser==5.2.1
Markdown==2.6.11
psycopg2==2.7.4
Pillow==5.1.0
versiontools==1.9.1
statsd==3.2.2
pep8==1.7.1
pyflakes==2.0.0
mccabe==0.6.1
configparser==3.5.0
pycodestyle==2.3.1 # pyup: <2.4.0
flake8==3.5.0
tornado==5.0.2
beautifulsoup4==4.6.0
cssselect==1.0.3
lxml==4.2.1
fuzzywuzzy==0.16.0
sure==1.4.11
traitlets==4.3.2  # ipython
ipython_genutils==0.2.0  # ipython
decorator==4.2.1  # ipython
ptyprocess==0.5.2  # ipython
pexpect==4.6.0  # ipython
pickleshare==0.7.4  # ipython
simplegeneric==0.8.1  # ipython
path.py==11.0.1  # ipython
wcwidth==0.1.7  # ipython
prompt_toolkit==1.0.15  # ipython
pygments==2.2.0  # ipython
scandir==1.7  # ipython
backcall==0.1.0  # ipython
ipython==6.4.0
ipdb==0.11
rdflib==4.2.2
coverage==4.5.1
asn1crypto==0.24.0  # cryptography
cryptography==2.2.2
urllib3==1.22  # requests
chardet==3.0.4  # requests
idna==2.6  # requests
certifi==2018.4.16  # requests
requests==2.18.4
logilab-common==1.4.1
logilab-astng==0.24.3
astroid==1.6.4
pylint==1.9.1
six==1.11.0
ipaddress==1.0.21  # faker
faker==0.8.15  # factory_boy
factory_boy==2.11.1
ldap3==2.5
sqlparse==0.2.4
boto==2.48.0
python-dateutil==2.7.3

# django libraries
djangowind==1.1.0
django-indexer==0.3.0
django-templatetag-sugar==1.0
django-paging==0.2.5
django-annoying==0.10.4
django-appconf==1.0.2
rcssmin==1.0.6  # django-compressor
rjsmin==1.0.12  # django-compressor
django-compressor==2.2
django-statsd-mozilla==0.4.0
raven==6.9.0
django-bootstrap4==0.0.6
django-debug-toolbar==1.9.1
django-waffle==0.14.0
django-jenkins==0.110.0
django-smoketest==1.1.0
typing==3.6.4  # For django-extensions
django-extensions==2.0.7
django-stagingcontext==0.1.0
django-ga-context==0.1.0
django-global-context==0.1.1
django-registration-redux==2.4
django-markwhat==1.6.0
gunicorn==19.8.1
django-infranil==1.1.0
django-flatblocks==0.9.4
django-storages-redux==1.3.3
django-cacheds3storage==0.1.2

djangorestframework==3.8.2
Willow==1.1
text-unidecode==1.2  # for faker
html5lib==1.0.1

ccnmtlsettings==1.4.0

oauthlib==2.1.0
oauth2==1.9.0.post1
pylti==0.6.0
nameparser==0.5.6
django-lti-provider==0.3.3

django-braces==1.13.0
django-smtp-ssl==1.0
django-contact-us==0.4.1

pbr==4.0.3  # bandit
PyYAML==3.12  # bandit
stevedore==1.28.0  # bandit
bandit==1.4.0
